+++ multiego/io.py
