+++ multiego/topology.py
