+++ multiego/write_output.py
